Project: Theme Template 5.0.0 Release 5.0.0: Introduce Substrate (MILESTONE RELEASE)
This commit will now be marked the bare minimum to pass current
Substratum versions that will be pushed to the Play Store.

Reverting of this commit will cause warnings from Substratum to inform
the user to inform the themer to upgrade their templates.

Themers: I have added a LuckyPatcher super enforcement library that
detects multiple levels of LuckyPatcher and an un-named masking tool.
It is highly recommended to keep this ENABLED, however if you prefer to
disable it, please follow the THIRD step on the last SECTION OF THE
README.

# Conflicts:
#	app/build.gradle
#	app/src/main/AndroidManifest.xml