Release 7.0.0: Completely rewrite the SubstratumThemeTemplate
From now on, themers should ONLY interact with the ThemerConstants.java
file, as that contains the most simple and easiest ways to
enable/disable certain features of antipiracy.

Themers can also add a Theme Changelog by changing the String-Array in
theme_strings.xml. If the themer does not want to have a changelog in
their theme, simply delete the whole entry.

Substrate now only functions as a simple class to launch the theme in
Substratum and does not function on any level of AntiPiracy. All
AntiPiracy package-aware checks are now located in ThemerConstants.java!
If the themer wants to be a big brother to a user and block say,
Hangouts, all they have to do is add "com.google.android.talk" into
BLACKLISTED_APPLICATIONS.

As with this version, this is the newest version of the template and
thus, there's a brand new hero image to reflect the version change!

This also fixes the bad dialog style when Theme Ready Gapps are
requested!